resolve by host label
