+++ README.md
@@ -14,12 +14,14 @@ Syntax
     docker [DOCKER_ENDPOINT] {
         domain DOMAIN_NAME
         network_aliases DOCKER_NETWORK
+        label LABEL
     }
 
 
 * `DOCKER_ENDPOINT`: the path to the docker socket. If unspecified, defaults to `unix:///var/run/docker.sock`. It can also be TCP socket, such as `tcp://127.0.0.1:999`.
 * `DOMAIN_NAME`: the name of the domain you want your containers to be part of. e.g., when `DOMAIN_NAME` is `docker.local`, your `mysql-0` container will be assigned the domain name: `mysql-0.docker.local`.
-* `DOCKER_NETWORK`: the name of the docker network
+* `DOCKER_NETWORK`: the name of the docker network. Resolve by network aliases as hosts (like internal docker dns resolve host by aliases whole network)
+* `LABEL`: label of resolving host (by default equals ```coredns.dockerdiscovery.host```)
 
 How To Build
 ------------
@@ -34,6 +36,11 @@ Alternatively, you can use the following manual steps:
 4. `go generate`
 5. `make`
 
+Alternatively, run insider docker container
+
+    docker build -t coredns-dockerdiscovery .
+    docker run --rm -v ${HOME}/Corefile:/coredns/Corefile -v /var/run/docker.sock:/var/run/docker.sock -p 15353:15353/udp coredns-dockerdiscovery -conf /coredns/Corefile
+
 Example
 -------
 
@@ -45,34 +52,6 @@ Example
         }
         log
     }
-
-    my-project.loc:15353 {
-        docker unix:///var/run/docker.sock {
-            network_aliases my_project_network
-        }
-        log
-    }
-    
-Create my_project_network and container
- 
-    docker create network my_project_network
-    docker run --network my_project_network --alias postgres.my-project.loc postgres
-
-Example for `docker-compose.yml`:
-
-    version: "3.1"
-
-    services:
-      postgresql:
-        image: ...
-        networks:
-          default:
-            aliases:
-              - postgres.my-project.loc
-    networks:
-      default:
-        external:
-          name: my_project_network
 
 Start CoreDNS:
 
@@ -89,7 +68,7 @@ Start a docker container:
     $ docker run -d --hostname alpha alpine sleep 1000
     78c2a06ef2a9b63df857b7985468f7310bba0d9ea4d0d2629343aff4fd171861
 
-Use CoreDNS as your resolver to resovle the `alpha.docker.local`:
+Use CoreDNS as your resolver to resolve the `alpha.docker.local`:
 
     $ dig @localhost -p 15353 alpha.docker.local
 
@@ -139,11 +118,42 @@ Stop the docker container will remove the DNS entry for `alpha.docker.local`:
     ;; WHEN: Thu Apr 26 22:41:38 EDT 2018
     ;; MSG SIZE  rcvd: 47
 
-Can resolve container by aliases
-           
-    $ dig @localhost -p 15353 postgres.my-project.loc
-    
-Alternatively, run insider docker container
+Can resolve container by network aliases
+     
+`Corefile`:
 
-    docker build -t coredns-dockerdiscovery .
-    docker run --rm -v ${PWD}/../Corefile:/coredns-dockerdiscovery/Corefile -v /var/run/docker.sock:/var/run/docker.sock -p 15353:15353/udp coredns-dockerdiscovery -conf /coredns-dockerdiscovery/Corefile
+    my-project.loc:15353 {
+        docker unix:///var/run/docker.sock {
+            network_aliases my_project_network
+        }
+        log
+    } 
+  
+Create my_project_network and container
+ 
+    docker create network my_project_network
+    docker run --network my_project_network --alias postgres.my-project.loc postgres
+
+Or example for `docker-compose.yml`:
+
+    version: "3.7"
+
+    services:
+      postgres:
+        image: postgres:latest
+        networks:
+          default:
+            aliases:
+              - postgres.my-project.loc
+    networks:
+      default:
+        external:
+          name: my_project_network  
+       
+Check
+       
+    $ dig @localhost -p 15353 postgres.my-project.loc
+
+Will resolve by label as ```nginx.loc```
+
+    docker run --rm --label=coredns.dockerdiscovery.host=nginx.loc nginx
